feat: future::now_or_never
## Description

This is almost exactly the same as
futures::future::FutureExt::now_or_never but as a function.  We can
not extend the FutureExt trait without duplicating it.  It also seems
like more the style of futures-lite.

## Breaking Changes

None

## Notes & open questions

I'm not sure if we should be adding our own implementations.  I
initially just created this function where I needed it, but upon
learning that futures has such a utility I thought it would be nice to
add here.  On the other hand if we only use this in one place maybe
that's not enough motivation to add it.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.

---------

Co-authored-by: dignifiedquire <me@dignifiedquire.com>